# src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "급식한눈 내부 API",
    "version": "1.0.0",
    "description": "프론트엔드와 백엔드 사이의 학교 검색 및 중식 조회 계약"
  },
  "paths": {
    "/api/schools": {
      "get": {
        "operationId": "searchSchools",
        "summary": "학교 이름 검색",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "minLength": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "학교 검색 결과",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SchoolSearchResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/UpstreamError" },
          "504": { "$ref": "#/components/responses/UpstreamError" }
        }
      }
    },
    "/api/schools/{educationOfficeCode}/{schoolCode}/meals": {
      "get": {
        "operationId": "getMeals",
        "summary": "학교 중식 조회",
        "parameters": [
          {
            "name": "educationOfficeCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "pattern": "^[A-Za-z0-9]+$" }
          },
          {
            "name": "schoolCode",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "pattern": "^[A-Za-z0-9]+$" }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "mealType",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "const": "lunch" }
          }
        ],
        "responses": {
          "200": {
            "description": "날짜순 중식 목록",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MealsResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": {
            "description": "학교를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "422": { "$ref": "#/components/responses/BadRequest" },
          "502": { "$ref": "#/components/responses/UpstreamError" },
          "503": { "$ref": "#/components/responses/UpstreamError" },
          "504": { "$ref": "#/components/responses/UpstreamError" }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "health",
        "summary": "서비스 상태 확인",
        "responses": {
          "200": {
            "description": "정상",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string", "const": "ok" } },
                  "required": ["status"]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "School": {
        "type": "object",
        "properties": {
          "educationOfficeCode": { "type": "string" },
          "schoolCode": { "type": "string" },
          "name": { "type": "string" },
          "region": { "type": "string" },
          "address": { "type": "string" }
        },
        "required": ["educationOfficeCode", "schoolCode", "name", "region", "address"],
        "additionalProperties": false
      },
      "SchoolSearchResponse": {
        "type": "object",
        "properties": {
          "schools": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/School" }
          }
        },
        "required": ["schools"],
        "additionalProperties": false
      },
      "Meal": {
        "type": "object",
        "properties": {
          "date": { "type": "string", "format": "date" },
          "mealType": { "type": "string", "const": "lunch" },
          "menu": { "type": "array", "items": { "type": "string" } },
          "calories": { "type": "string" },
          "nutrition": {
            "type": "object",
            "additionalProperties": { "type": "string" }
          },
          "origin": { "type": "string" }
        },
        "required": ["date", "mealType", "menu"],
        "additionalProperties": false
      },
      "MealsResponse": {
        "type": "object",
        "properties": {
          "school": { "$ref": "#/components/schemas/SchoolSummary" },
          "from": { "type": "string", "format": "date" },
          "to": { "type": "string", "format": "date" },
          "meals": { "type": "array", "items": { "$ref": "#/components/schemas/Meal" } }
        },
        "required": ["school", "from", "to", "meals"],
        "additionalProperties": false
      },
      "SchoolSummary": {
        "type": "object",
        "properties": {
          "educationOfficeCode": { "type": "string" },
          "schoolCode": { "type": "string" },
          "name": { "type": "string" }
        },
        "required": ["educationOfficeCode", "schoolCode", "name"],
        "additionalProperties": false
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "requestId": { "type": "string" }
        },
        "required": ["code", "message", "requestId"],
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "properties": { "error": { "$ref": "#/components/schemas/Error" } },
        "required": ["error"],
        "additionalProperties": false
      }
    },
    "responses": {
      "BadRequest": {
        "description": "잘못된 입력",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "UpstreamError": {
        "description": "NEIS 연동 실패",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  }
}